Refator: first refactoring
