+++ src/main/java/com/example/ddo_auth/auth/dto/EmailLoginRequest.java
@@ -1,0 +1,21 @@
+package com.example.ddo_auth.auth.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Builder;
+import lombok.AllArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class EmailLoginRequest {
+    @NotBlank(message = "이메일은 필수 입력값입니다")
+    @Email(message = "이메일 형식이 올바르지 않습니다")
+    private String email;
+
+    @NotBlank(message = "비밀번호는 필수 입력값입니다")
+    private String password;
+}
