+++ src/main/java/com/example/ddo_auth/auth/dto/TokenResponse.java
@@ -1,0 +1,11 @@
+package com.example.ddo_auth.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenResponse {
+    private String accessToken;
+    private String refreshToken;
+}
