+++ src/main/java/com/example/ddo_auth/auth/security/CustomUserDetailsService.java
@@ -1,0 +1,24 @@
+package com.example.ddo_auth.auth.security;
+
+import com.example.ddo_auth.auth.entity.User;
+import com.example.ddo_auth.auth.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    //유저 조회
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("해당 이메일을 가진 사용자를 찾을 수 없습니다."));
+        return new CustomUserDetails(user);
+    }
+}
