--- src/main/java/com/example/ddo_auth/config/CustomAuthorizationRequestResolver.java
@@ -1,45 +1,0 @@
-package com.example.ddo_auth.config;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizationRequestResolver;
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizationRequestResolver;
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.stereotype.Component;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-public class CustomAuthorizationRequestResolver implements OAuth2AuthorizationRequestResolver {
-
-    private final DefaultOAuth2AuthorizationRequestResolver defaultResolver;
-
-    public CustomAuthorizationRequestResolver(ClientRegistrationRepository repo) {
-        this.defaultResolver = new DefaultOAuth2AuthorizationRequestResolver(repo, "/oauth2/authorization");
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest resolve(HttpServletRequest request) {
-        OAuth2AuthorizationRequest req = defaultResolver.resolve(request);
-        return customizeAuthorizationRequest(req);
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest resolve(HttpServletRequest request, String clientRegistrationId) {
-        OAuth2AuthorizationRequest req = defaultResolver.resolve(request, clientRegistrationId);
-        return customizeAuthorizationRequest(req);
-    }
-
-    private OAuth2AuthorizationRequest customizeAuthorizationRequest(OAuth2AuthorizationRequest req) {
-        if (req == null) return null;
-
-        Map<String, Object> extraParams = new HashMap<>(req.getAdditionalParameters());
-        extraParams.put("access_type", "offline");
-        extraParams.put("prompt", "consent");
-
-        return OAuth2AuthorizationRequest.from(req)
-                .additionalParameters(extraParams)
-                .build();
-    }
-}
