--- src/main/java/com/example/ddo_auth/domain/User.java
@@ -1,37 +1,0 @@
-package com.example.ddo_auth.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "user_id", columnDefinition = "BINARY(16)")
-    private UUID userId;
-
-    @Column(nullable = false, unique = true, length = 255)
-    private String email;
-
-    @Column(nullable = false, length = 255)
-    private String password;
-
-    @Column(length = 100, columnDefinition = "VARCHAR(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci")
-    private String name;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Provider provider;
-
-    public enum Provider {
-        LOCAL, GOOGLE
-    }
-}
