--- src/main/java/com/example/ddo_auth/jwt/JwtTokenProvider.java
@@ -1,91 +1,0 @@
-package com.example.ddo_auth.jwt;
-
-import com.example.ddo_auth.security.CustomUserDetailsService;
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-
-@Component
-@RequiredArgsConstructor
-public class JwtTokenProvider {
-
-    @Value("${jwt.secret}")
-    private String secretKey;
-
-    private Key key;
-
-    @Value("${jwt.access-token-expiration}")
-    private long accessTokenExpiration;
-
-    @Value("${jwt.refresh-token-expiration}")
-    private long refreshTokenExpiration;
-
-    @PostConstruct
-    protected void init() {
-        this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public String createAccessToken(String userEmail) {
-        return createToken(userEmail, accessTokenExpiration);
-    }
-
-    public String createRefreshToken(String userEmail) {
-        return createToken(userEmail, refreshTokenExpiration);
-    }
-
-    private String createToken(String userEmail, long validityInMillis) {
-        Date now = new Date();
-        Date expiry = new Date(now.getTime() + validityInMillis);
-
-        return Jwts.builder()
-                .setSubject(userEmail)
-                .setIssuedAt(now)
-                .setExpiration(expiry)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    public String getUserEmailFromToken(String token) {
-        Claims claims = Jwts.parserBuilder().setSigningKey(key).build()
-                .parseClaimsJws(token).getBody();
-        return claims.getSubject();
-    }
-
-
-
-    public String resolveToken(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7); // "Bearer " 이후 문자열 반환
-        }
-        return null;
-    }
-
-    private final CustomUserDetailsService userDetailsService;
-
-    public Authentication getAuthentication(String token) {
-        String email = getUserEmailFromToken(token);
-        UserDetails userDetails = userDetailsService.loadUserByUsername(email);
-        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
-    }
-}
