--- src/main/java/com/example/ddo_auth/repository/RefreshTokenRepository.java
@@ -1,7 +1,0 @@
-package com.example.ddo_auth.repository;
-
-import com.example.ddo_auth.domain.RefreshToken;
-import org.springframework.data.repository.CrudRepository;
-
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-}
